Add removeRow() method to board, fixes DASH-102
