+++ integration-tests/src/test/java/com/vaadin/addon/board/testbenchtests/RemoveComponentIT.java
@@ -1,13 +1,12 @@
 package com.vaadin.addon.board.testbenchtests;
 
-import java.util.List;
-
 import org.junit.Assert;
 import org.junit.Test;
 import org.openqa.selenium.WebElement;
 
 import com.vaadin.addon.board.testUI.RemoveComponentUI;
 import com.vaadin.testbench.By;
+import com.vaadin.testbench.elements.ButtonElement;
 
 public class RemoveComponentIT extends AbstractParallelTest {
 
@@ -17,11 +16,13 @@ public class RemoveComponentIT extends AbstractParallelTest {
     }
 
     @Test
-    public void basicLayout_removeComponentFromRow_removedComponentsNotShown() {
+    public void basicLayout_boardTabletSize_twoRowsAndTwoItemsInRow() {
         WebElement board =getDriver().findElement(By.tagName("vaadin-board"));
+        ButtonElement btn1 = $(ButtonElement.class).caption("Button 1").first();
+        ButtonElement btn2 = $(ButtonElement.class).caption("Button 3").first();
 
-        List<WebElement> children = board.findElements(By.xpath("//vaadin-board/vaadin-board-row/*"));
-        Assert.assertEquals("Board should have 2 children", 2, children.size());
+        Assert.assertEquals("Btn1 should have Y == 0" , board.getLocation().getY(), btn1.getLocation().getY());
+        Assert.assertEquals("Btn2 should have Y == 0", board.getLocation().getY(), btn2.getLocation().getY());
 
     }
